requirements.txt
Django==2.2.18
django-environ==0.4.5
djangorestframework==3.12.2
gunicorn==20.0.4
pytz==2021.1
scanapi==2.1.0
sqlparse==0.4.1
